# docs/swagger.yaml
basePath: /
definitions:
...
  models.Table:
    properties:
      created_at:
        type: string
      id:
        type: string
      number_of_guests:
        type: integer
      table_id:
        type: string
      table_number:
        type: integer
      updated_at:
        type: string
    required:
    - number_of_guests
    - table_number
    type: object
  models.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        maxLength: 100
        minLength: 2
        type: string
      id:
        type: string
      last_name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      refresh_token:
        type: string
      token:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - phone
    type: object
host: localhost:8000
info:
  contact: {}
  description: CRUP api's for go backend
  title: go-backend
  version: "1.0"
paths:
...
  /tables:
    get:
      description: Returns an array of tables placed from table collectios in restaurent
        database.
      produces:
      - application/json
      responses:
        "200":
          description: 'slice of tables '
          schema:
            items:
              $ref: '#/definitions/models.Table'
            type: array
      security:
      - Security.require(true): []
      summary: Returns slice of tables
      tags:
      - table
    post:
      consumes:
      - application/json
      description: Creates a new table resource on the server
      parameters:
      - description: Table object
        in: body
        name: table
        required: true
        schema:
          $ref: '#/definitions/models.Table'
      produces:
      - application/json
      responses:
        "201":
          description: New table created
          schema:
            $ref: '#/definitions/models.Table'
        "500":
          description: Internal Server Error while creating a new table
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Creates a table  resource
      tags:
      - table
  /tables/{table_id}:
    get:
      consumes:
      - application/json
      description: Retrieves a table with specific table id from the tables collection
      parameters:
      - description: table_id
        in: path
        name: table_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: table details of specific userid
          schema:
            $ref: '#/definitions/models.Table'
        "500":
          description: Internal Server Error in mongodb
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Retrieves a table  with specific table id
      tags:
      - table
    patch:
      consumes:
      - application/json
      description: Updates an existing table resource in the table collection
      parameters:
      - description: ID of the table resource to update
        in: path
        name: table_id
        required: true
        type: string
      - description: Table object
        in: body
        name: table
        required: true
        schema:
          $ref: '#/definitions/models.Table'
      produces:
      - application/json
      responses:
        "200":
          description: table got updated with new body
          schema:
            $ref: '#/definitions/models.Table'
        "404":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error while updating table
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Updates an table resource
      tags:
      - table
  /user/login:
    post:
      consumes:
      - application/json
      description: User gets logged in
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: 'User logged in '
          schema:
            $ref: '#/definitions/models.User'
        "500":
          description: Internal Server Error while logging  a new user
          schema:
            type: string
      summary: Creates a new login for a user
      tags:
      - user
...
  /users/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieves a user with specific user id from the users collection
      parameters:
      - description: user_id
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Details of user with specific user_id
          schema:
            $ref: '#/definitions/models.User'
        "500":
          description: Internal Server Error in mongodb
          schema:
            type: string
      summary: Retrieves a user with specific user id
      tags:
      - user
  /users/signup:
    post:
      consumes:
      - application/json
      description: Creates a new users resource on the server
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: New user created
          schema:
            $ref: '#/definitions/models.User'
        "500":
          description: Internal Server Error while creating a new user
          schema:
            type: string
      summary: Creates a new resource
      tags:
      - user
securityDefinitions:
  "":
    description: JWT token for authentication
    in: header
    name: token
    type: apiKey
swagger: "2.0"